refactor: Move launching NS functions to main.rs
Part of #329
